requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
distlib==0.3.0
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
filelock==3.0.12
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pbr==5.4.4
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
stevedore==1.32.0
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==20.0.7
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wrapt==1.11.2
zipp==3.0.0
